refactor(transpier): rename _scope to _aliasing
